refactor(governance): 拆分 git-history 分支菜单与面板类型
